rand shuffle kinda working, except for Taillard's format and changing time limit (default is 3 min)

+++ jobshop/input_output.h
@@ -7,4 +7,4 @@
 
 void read_t(std::string filename, int* machines_c, int* jobs_c, int**& machines, int**& jobs, int max_jobs);
 void read_b(std::string filename, int* machines_c, int* jobs_c, int**& machines, int**& jobs, int max_jobs);
-void write_to_file(std::string filename, int machines_c, int jobs_c, int**& times);
+void write_to_file(std::string filename, int machines_c, int jobs_c, int64_t**& times);
